+++ src/apps/vectorwar/vectorwar.h
@@ -19,8 +19,8 @@ enum VectorWarInputs {
    INPUT_BOMB              = (1 << 5),
 };
 
-void VectorWar_Init(HWND hwnd, int localport, int num_players, GGPOPlayer *players, int num_spectators);
-void VectorWar_InitSpectator(HWND hwnd, int localport, int num_players, char *host_ip, int host_port);
+void VectorWar_Init(HWND hwnd, unsigned short localport, int num_players, GGPOPlayer *players, int num_spectators);
+void VectorWar_InitSpectator(HWND hwnd, unsigned short localport, int num_players, char *host_ip, unsigned short host_port);
 void VectorWar_DrawCurrentFrame();
 void VectorWar_AdvanceFrame(int inputs[], int disconnect_flags);
 void VectorWar_RunFrame(HWND hwnd);
