+++ src/include/ggponet.h
@@ -404,9 +404,9 @@ GGPO_API GGPOErrorCode __cdecl ggpo_start_spectating(GGPOSession **session,
                                                      const char *game,
                                                      int num_players,
                                                      int input_size,
-                                                     unsigned int local_port,
+                                                     unsigned short local_port,
                                                      char *host_ip,
-                                                     unsigned int host_port);
+                                                     unsigned short host_port);
 
 /*
  * ggpo_close_session --
